backend/requirements.txt
django>=2.2.6
djangorestframework
drf-generators
